🍒 Move logic of showing categories popup menu to `TaskUtils`
