+++ app/src/main/java/com/example/sunrise/helpers/TaskCreationHelper.java
@@ -3,16 +3,13 @@ package com.example.sunrise.helpers;
 import android.content.Context;
 import android.content.res.ColorStateList;
 import android.util.Log;
-import android.util.Pair;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.Button;
-import android.widget.PopupMenu;
 
 import androidx.annotation.NonNull;
 
 import com.example.sunrise.R;
-import com.example.sunrise.models.Category;
 import com.example.sunrise.models.Tag;
 import com.example.sunrise.models.Task;
 import com.example.sunrise.services.CategoryService;
@@ -30,9 +27,7 @@ import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.ValueEventListener;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 
 public class TaskCreationHelper {
@@ -73,7 +68,7 @@ public class TaskCreationHelper {
         priorityChip.setOnClickListener(view -> TaskUtils.showPriorityPopupMenu(context, view, priorityChip));
 
         categoryChip = bottomSheetContentView.findViewById(R.id.category);
-        categoryChip.setOnClickListener(view -> showCategoriesPopupMenu(view, categoryChip));
+        categoryChip.setOnClickListener(view -> TaskUtils.showCategoriesPopupMenu(context, view, categoryChip, categoryService, this::onCategorySelected));
 
         // Tags row
         tagChips = bottomSheetContentView.findViewById(R.id.tagChips);
@@ -82,65 +77,12 @@ public class TaskCreationHelper {
         createBtn.setOnClickListener(this::createTask);
     }
 
-    private void showCategoriesPopupMenu(View v, Chip categoryChip) {
-        PopupMenu popup = new PopupMenu(context, categoryChip);
+    private void onCategorySelected(String selectedCategoryId, String selectedCategoryName) {
+        // Store it like class var
+        this.selectedCategoryId = selectedCategoryId;
 
-        // Retrieve categories from the database
-        retrieveCategoriesFromDatabase(categoryService, categoryIdMap -> {
-            // Add categories to the popup menu
-            for (Map.Entry<Integer, Pair<String, String>> entry : categoryIdMap.entrySet()) {
-                popup.getMenu().add(0, entry.getKey(), 0, entry.getValue().second);
-            }
-
-            // Show popupMenu after all asynchronous stuff is done
-            popup.show();
-
-            popup.setOnMenuItemClickListener(item -> {
-                // Handle category selection
-                int categoryIdHashCode = item.getItemId();
-
-                Pair<String, String> selectedCategoryInfo = categoryIdMap.get(categoryIdHashCode); // Retrieve Pair using hash code
-                assert selectedCategoryInfo != null;
-
-                // Store it like class var
-                this.selectedCategoryId = selectedCategoryInfo.first;
-
-                // Set selected category's title to chip
-                categoryChip.setText(selectedCategoryInfo.second);
-                return true;
-            });
-        });
-    }
-
-    private void retrieveCategoriesFromDatabase(CategoryService categoryService, TaskCreationHelper.CategoriesRetrievedCallback callback) {
-        Map<Integer, Pair<String, String>> categoryIdMap = new HashMap<>(); // Map to store category IDs and their hash codes
-        categoryService.getCategories(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
-                // Add categories to the categoryIdMap
-                for (DataSnapshot snapshot : dataSnapshot.getChildren()) {
-                    Category category = snapshot.getValue(Category.class);
-                    if (category != null) {
-                        String categoryId = category.getCategoryId();
-                        String categoryName = category.getTitle();
-                        int categoryIdHashCode = categoryId.hashCode();  // Convert categoryId to hashcode cause PopupMenu add item method needs int as id
-                        categoryIdMap.put(categoryIdHashCode, new Pair<>(categoryId, categoryName)); // Store the relationship between hash code and Pair of categoryId/categoryName
-                    }
-                }
-
-                // Invoke the callback with passed categoryIdMap
-                callback.onCategoriesRetrieved(categoryIdMap);
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError error) {
-                Log.e("CategoriesFragment", "Getting tags failed");
-            }
-        });
-    }
-
-    interface CategoriesRetrievedCallback {
-        void onCategoriesRetrieved(Map<Integer, Pair<String, String>> tagIdMap);
+        // Set selected category's title to chip
+        categoryChip.setText(selectedCategoryName);
     }
 
     private void populateTagChips() {
